+++ NEWS.md
@@ -4,6 +4,9 @@
 ## New features and improvements
 
 * ADSL update to have all factor levels of RACE and ETHNIC
+* Fixed ADCM to properly handle uncoded terms
+* Fixed ADLB to use "BEFORE TREATMENT" for ATPT instead of "00:00"
+
 
 ## Changes